examples/calling-apis/llama-index-examples/pyproject.toml
[tool.poetry]
name = "llama-index-examples"
version = "0.1.0"
description = ""
authors = ["Auth0 <support@auth0.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.scripts]
start = "scripts.start:main"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
flask = { extras = ["async"], version = "^3.1.0" }
requests = "^2.32.3"
hypercorn = "^0.17.3"
llama-index = "^0.12.33"
llama-index-agent-openai = "^0.4.6"
auth0-server-python = "^1.0.0b4"
auth0-ai-llamaindex = { path = "../../../packages/auth0-ai-llamaindex", develop = true }
slack-sdk = "^3.35.0"
pygithub = "^2.6.1"
asgiref = "^3.8.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
